+++ maze_game/maze_game/Shapes.cs
@@ -29,103 +29,9 @@ namespace maze_game
 
         }
     }
-
-    public class Box : Shapes
-    {
-        public override void Draw(bool isDialogue = false)
-        {
-            Console.ForegroundColor = ForegroundColour;
-            Console.BackgroundColor = BackgroundColour;
-
-            for(int width = 0; width < ShapeSize.Width; width++)
-            {
-                for(int height = 0; height < ShapeSize.Height; height++)
-                {
-                    Console.SetCursorPosition(ShapePosition.X + width, ShapePosition.Y + height);
-                    if(width == 0 && height == 0)
-                    {
-                        Console.Write("╔");
-                    }
-                    else if (width == ShapeSize.Width - 1 && height == 0)
-                    {
-                        Console.Write("╗");
-                    }
-                    else if (width == 0 && height == ShapeSize.Height - 1)
-                    {
-                        Console.Write("╚");
-                    }
-                    else if (width == ShapeSize.Width - 1 && height == ShapeSize.Height - 1)
-                    {
-                        Console.Write("╝");
-                    }
-                    else if( (width == 0 || width == ShapeSize.Width - 1) && (height > 0 && height < ShapeSize.Height - 1))
-                    {
-                        if (isDialogue)
-                        {
-                            if(height == 2)
-                            {
-                                if (width == 0)
-                                {
-                                    Console.Write("╠");
-                                }
-                                else if (width == ShapeSize.Width - 1)
-                                {
-                                    Console.Write("╣");
-                                }
-                            }
-                            else
-                            {
-                                Console.Write("|");
-                            }
-                        }
-                        else
-                        {
-                            Console.Write("║");
-                        }
-                    }
-                    else if ((width > 0 && width < ShapeSize.Width - 1) && (height == 0 || height == ShapeSize.Height - 1))
-                    {
-                        Console.Write("=");
-                    }
-                    else
-                    {
-                        Console.Write(" ");
-                    }
-                }
-            }
-        }
-    }
-
-    public class DialogueBox : Shapes
-    {
-        Box textbox = new Box();
-        string Title;
-
-        public override void Create(Point position, Size size, ConsoleColor foregroundColour, ConsoleColor backgorundColour, string title = "")
-        {
-            Title = title;
-
-            textbox.Create(position, size, foregroundColour, backgorundColour);
-
-            if (Title.Length >= textbox.ShapeSize.Width)
-            {
-                textbox.ShapeSize = new Size(Title.Length + 4, textbox.ShapeSize.Height);
-            }
-        }
-
-        public override void Draw(bool isDialogue = true)
-        {
-            textbox.Draw(isDialogue);
-
-            Console.SetCursorPosition(textbox.ShapePosition.X + 1, textbox.ShapePosition.Y + 2);
-
-            for(int width = 0; width < textbox.ShapeSize.Width - 2; width++)
-            {
-                Console.Write("=");
-            }
-
-            Console.SetCursorPosition(textbox.ShapePosition.X + 2, textbox.ShapePosition.Y + 1);
-            Console.Write(Title);
-        }
-    }
 }
+
+
+
+
+
